requirements.txt
openai>=1.20.0
PyPDF2>=3.0.0
pandas>=2.0.0
tqdm>=4.65.0
argparse>=1.4.0
concurrent-futures>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
umap-learn>=0.5.3
plotly>=5.13.0
dash>=2.9.0
hdbscan>=0.8.29
